Task1 + Refactoring
