+++ src/teacher/lesson_2/homework/level_1_intern/Task_1.txt
@@ -3,9 +3,11 @@
 
 1. Создать класс с нужным именем;
 2. Создать описать main метод;
-3. Воспользуйтесь специальным классом Scanner и его методом nextInt() для получения ввода с клавиатуры
+3. Воспользуйтесь специальным классом Scanner и его методом nextInt()
+ для получения ввода с клавиатуры
           Scanner scanner = new Scanner(System.in);
           int firstNumber = scanner.nextInt();
 4. Запишите полученные результаты в переменные;
-5. При помощи System.out.println() выведите на консоль полученные результаты сложения, вычитания, умножения и деления.
+5. При помощи System.out.println() выведите на консоль полученные результаты
+ сложения, вычитания, умножения и деления.
  (каждый результат на новой строке отдельно).
